Implement generate_x dependencies

+++ lib/include/PoissonRegressionTrainer.hpp
@@ -13,7 +13,7 @@ class PoissonRegressionTrainer {
         PoissonRegressionTrainer();
         PoissonRegressionTrainer(DataFramePosRegTransformer* transforms);
         PoissonRegressionModelData get_poisson_regression_model_data(ULDataFrame df, std::string y_col_name);
-        std::vector<std::vector<double>> generate_x(ULDataFrame df, PoissonRegressionModelData model_data);
+        std::vector<std::vector<unsigned int>> generate_x(ULDataFrame df, PoissonRegressionModelData model_data);
     private:
         DataFramePosRegTransformer* transforms;
 };
